+++ src/C/context.jl
@@ -309,8 +309,8 @@ function init_context()
         error("Cannot parse version from version string: $(repr(verstr))")
     end
     CTX.version = VersionNumber(vermatch.match)
-    v"3.9" ≤ CTX.version < v"4" || error(
-        "Only Python 3.9+ is supported, this is Python $(CTX.version) at $(CTX.exe_path===missing ? "unknown location" : CTX.exe_path).",
+    v"3.10" ≤ CTX.version < v"4" || error(
+        "Only Python 3.10+ is supported, this is Python $(CTX.version) at $(CTX.exe_path===missing ? "unknown location" : CTX.exe_path).",
     )
 
     launch_on_main_thread(Threads.threadid()) # makes on_main_thread usable